Move FlexibleTransportMessageBox to its own file
